I am successfully deleting categories and posts I did some console.log in the login.js page because I notice I just get an error for a wrond password etc. so I started hunting that issue downfor a second but soon got back on the topic of deleting blogPost and categories as per this feature I did work in the deleteCategory.js api page to handle categories that have associated blogPosts as I do not want to delete a category if it still has blogPost associated in the deleteBlogPosts.js api route I added logic to pull post data from a given category when a post is deleted Im calling this feature done commit before develop mege and push

+++ topdev-blog/pages/api/deleteCategory.js
@@ -9,6 +9,10 @@ export default async function deleteCategory({body},res) {
         res.status(500).json({message:'problem with db connection'});
     };
     try{
+        const categoryToDelete = await Category.findOne({_id: body.category});
+        if(categoryToDelete.blogPosts.length) {
+            return res.status(401).json({message:'this category has associated Blog Posts cannot delete'});
+        }
        const deletedCategory = await Category.findOneAndDelete({_id: body.category});
        res.status(200).json(deletedCategory);
     }catch(e){
